pyproject.toml
[project]
name = "personal-learning-coach"
version = "0.1.0"
description = "Closed-loop personal AI learning coach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
coach = "personal_learning_coach.coach:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/personal_learning_coach"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=personal_learning_coach --cov-report=term-missing"

[tool.coverage.run]
omit = [
    "*/coach.py",
    "*/scheduler_adapter.py",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]
